Use Thor for built-in middleware task
Currently we use both Thor and Rake for `bin/rails` commands.
We eventually want to get all the built-in tasks promoted to Thor Commands.
This migrates the `middleware` task to Thor.
