+++ railties/lib/rails/tasks.rb
@@ -6,7 +6,6 @@ require "rake"
 %w(
   framework
   log
-  middleware
   misc
   restart
   tmp
