--- railties/lib/rails/tasks/middleware.rake
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-desc "Print out your Rack middleware stack"
-task middleware: :environment do
-  Rails.configuration.middleware.each do |middleware|
-    puts "use #{middleware.inspect}"
-  end
-  puts "run #{Rails.application.class.name}.routes"
-end
